requirements.txt
pandas==1.2.2
python-dotenv==0.15.0
requests==2.25.1
